PBSmodelling-UG updated for v.2.20
Note: I notice that testWidgets.Rd was not updated to include the four new objects. I also suspect that the argument lists for the existing widgets in this same Rd file also need updating. I leave this up to Alex and Jon to decide. (BTW, I was too lazy to create this as an issue).