Allow multiple on-screen windows to have different bg colours (Issues 3 & 84)
The inability to use different background colours on multiple
on-screen windows stemmed from documented behaviour of tk_setPalette.
This function, which we call for new windows, is responsible for
deriving colours from the background colour, e.g., for highlights.
The problem: the function changes the colours of existing widgets in
addition to any future widgets when those existing widgets are
coloured with the default colour.

In this fix, we slightly adjust the colour of every widget as it is
created.  By making these small adjustments, widgets never use their
default colour, and subsequent calls to tk_setPalette don't change
their colours.  The drawback: we no longer honour user colour requests
exactly.